requirements.txt
requests
pyyaml
eth_utils